Moved charmEditor modules to lib
