+++ components/common/CharmEditor/components/ResizableIframe.tsx
@@ -7,49 +7,15 @@ import PreviewIcon from '@mui/icons-material/Preview';
 import VideoLibraryIcon from '@mui/icons-material/VideoLibrary';
 import { ListItem, Typography } from '@mui/material';
 import { Box } from '@mui/system';
+import { MAX_EMBED_WIDTH, MIN_EMBED_HEIGHT, MAX_EMBED_HEIGHT, VIDEO_ASPECT_RATIO, MIN_EMBED_WIDTH } from 'lib/embed/constants';
+import { extractEmbedLink } from 'lib/embed/extractEmbedLink';
 import { HTMLAttributes, useState, memo } from 'react';
 import BlockAligner from './BlockAligner';
 import IFrameSelector from './IFrameSelector';
 import Resizable from './Resizable';
 import VerticalResizer from './Resizable/VerticalResizer';
 
-export const MAX_EMBED_WIDTH = 700;
-export const MIN_EMBED_WIDTH = 100;
-export const MAX_EMBED_HEIGHT = 2500;
-export const MIN_EMBED_HEIGHT = 200;
-export const VIDEO_ASPECT_RATIO = 1.77;
-
 const name = 'iframe';
-
-export function extractEmbedLink (url: string) {
-  const isIframeEmbed = url.startsWith('<iframe ');
-  let embedUrl = url;
-
-  const isRegularYTLink = url.includes('youtube');
-  const isSharedYTLink = url.includes('youtu.be');
-  if (isRegularYTLink || isSharedYTLink) {
-    const { pathname, search } = new URL(url);
-    const urlSearchParams = new URLSearchParams(search);
-    if (isRegularYTLink) {
-      embedUrl = `https://www.youtube.com/embed/${urlSearchParams.get('v')}`;
-    }
-    else if (isSharedYTLink) {
-      embedUrl = `https://www.youtube.com/embed${pathname}`;
-    }
-    if (urlSearchParams.has('t')) {
-      embedUrl += `?start=${urlSearchParams.get('t')}`;
-    }
-  }
-
-  else if (isIframeEmbed) {
-    const indexOfSrc = url.indexOf('src');
-    const indexOfFirstQuote = url.indexOf('"', indexOfSrc);
-    const indexOfLastQuote = url.indexOf('"', indexOfFirstQuote + 1);
-    embedUrl = url.slice(indexOfFirstQuote + 1, indexOfLastQuote);
-  }
-
-  return embedUrl;
-}
 
 interface DispatchFn {
   (tr: Transaction): void;
