+++ lib/image/constants.ts
@@ -1,0 +1,2 @@
+export const MAX_IMAGE_WIDTH = 750;
+export const MIN_IMAGE_WIDTH = 100;
